+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 0.1.10 - 2025-08-26
 - Link sensor and button to a renamable device
 
+## Unreleased
+- Allow reconfiguring consumables from the UI including name, type, duration and expiry date override
+
 ## 0.1.8 - 2025-08-26
 - 10th Tries a charm?